More tests and fixes.  Pro tip: don't send a pointer to your allocator to your placement new function that takes a reference.

+++ RenderMode/Memory/LinearAllocator.cpp
@@ -30,7 +30,7 @@ size_t LinearAllocator::maxSize() {
 }
 
 size_t LinearAllocator::curSize() {
-    return cur - start;
+    return (size_t)(cur - start);
 }
